+++ back/server/controllers/group/createGroup.js
@@ -1,0 +1,1 @@
+export default function createGroup() {}
