+++ back/server/controllers/group/index.js
@@ -1,0 +1,1 @@
+export { default as createGroup } from "./createGroup.js";
