+++ back/server/controllers/user/index.js
@@ -1,0 +1,1 @@
+export { default as register } from "./register.js";
