+++ back/server/initializeServer.js
@@ -1,0 +1,12 @@
+const initializeServer = (app, port) =>
+  new Promise((resolve, reject) => {
+    const server = app.listen(port, () => {
+      resolve();
+    });
+
+    server.on("error", (error) => {
+      reject(error);
+    });
+  });
+
+export default initializeServer;
