+++ back/server/routes/index.js
@@ -1,0 +1,2 @@
+export { default as userRoutes } from "./user/index.js";
+export { default as groupRoutes } from "./group/index.js";
