+++ back/server/routes/user/index.js
@@ -1,0 +1,8 @@
+import express from "express";
+
+import { register } from "../../controllers/user/index.js";
+const router = express.Router();
+
+router.post("/register", register);
+
+export default router;
