+++ back/websocket/initializeWebsocket.js
@@ -1,0 +1,14 @@
+import { WebSocketServer } from "ws";
+
+const initializeWebsocket = (port) =>
+  new Promise((resolve, reject) => {
+    const server = new WebSocketServer({ port }, () => {
+      resolve();
+    });
+
+    server.on("error", (error) => {
+      reject(error);
+    });
+  });
+
+export default initializeWebsocket;
